Update image.py
Changed line 106 to include a +1 to make this sync with the rest of the code
changed line 114 'persist_frame' to 'frame_persist' .  This makes only the reel that changes on a next ball advance rather than have all of the player's score reels advance on the next ball.  For example with this change 1410 goes to 1510 on a score of 100 on the next ball rather than 1400 -> 1000 -> 1110 -> 1210 -> 1310 -> 1410 -> 1510.